store: Create a deployment with a plain connection
No need to have a fake entities::Connection
